Se agregan las funciones a archivos fuera del main para mejorar la visualizacion del codigo.
