+++ app/src/main/java/com/tecmilenio/booktrackerevidencia/AddBook.kt
@@ -1,0 +1,128 @@
+package com.tecmilenio.booktrackerevidencia
+
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
+import coil.compose.rememberAsyncImagePainter
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun AddBookDialog(
+    onDismiss: () -> Unit,
+    onBookAdded: (Book) -> Unit
+) {
+    var title by remember { mutableStateOf("") }
+    var author by remember { mutableStateOf("") }
+    var totalPages by remember { mutableStateOf("") }
+    var genre by remember { mutableStateOf("") }
+    var startDate by remember { mutableStateOf("") }
+    var imageUri by remember { mutableStateOf<String?>(null) }
+
+    var errorMessage by remember { mutableStateOf<String?>(null) } // Mensaje de error
+
+    val context = LocalContext.current
+    val launcher = rememberLauncherForActivityResult(ActivityResultContracts.GetContent()) { uri ->
+        uri?.let { imageUri = it.toString() }
+    }
+
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text("Agregar Libro") },
+        text = {
+            Column {
+                if (errorMessage != null) {
+                    Text(
+                        text = errorMessage!!,
+                        color = Color.Red,
+                        modifier = Modifier.padding(bottom = 10.dp)
+                    )
+                }
+
+                OutlinedTextField(
+                    value = title,
+                    onValueChange = { title = it },
+                    label = { Text("Título") },
+                    isError = title.isEmpty() && errorMessage != null // Resaltar campo vacío
+                )
+                OutlinedTextField(
+                    value = author,
+                    onValueChange = { author = it },
+                    label = { Text("Autor") },
+                    isError = author.isEmpty() && errorMessage != null
+                )
+                OutlinedTextField(
+                    value = totalPages,
+                    onValueChange = { totalPages = it },
+                    label = { Text("Páginas Totales") },
+                    isError = (totalPages.isEmpty() || totalPages.toIntOrNull() == null) && errorMessage != null
+                )
+                OutlinedTextField(
+                    value = genre,
+                    onValueChange = { genre = it },
+                    label = { Text("Género") },
+                    isError = genre.isEmpty() && errorMessage != null
+                )
+                OutlinedTextField(
+                    value = startDate,
+                    onValueChange = { startDate = it },
+                    label = { Text("Fecha de inicio") },
+                    isError = startDate.isEmpty() && errorMessage != null
+                )
+
+                Spacer(modifier = Modifier.height(10.dp))
+
+                Button(onClick = { launcher.launch("image/*") }) {
+                    Text("Seleccionar Imagen")
+                }
+
+                imageUri?.let {
+                    Image(
+                        painter = rememberAsyncImagePainter(it),
+                        contentDescription = "Imagen del libro",
+                        modifier = Modifier
+                            .size(100.dp)
+                            .padding(top = 10.dp)
+                    )
+                }
+            }
+        },
+        confirmButton = {
+            Button(onClick = {
+                if (title.isEmpty() || author.isEmpty() || totalPages.isEmpty() || genre.isEmpty() || startDate.isEmpty()) {
+                    errorMessage = "Todos los campos son obligatorios"
+                } else if (totalPages.toIntOrNull() == null) {
+                    errorMessage = "El número de páginas debe ser un valor válido"
+                } else {
+                    onBookAdded(Book(title, author, totalPages.toInt(), genre, startDate, imageUri))
+                    onDismiss()
+                }
+            }) {
+                Text("Agregar")
+            }
+        },
+        dismissButton = {
+            Button(onClick = onDismiss) {
+                Text("Cancelar")
+            }
+        }
+    )
+}
