+++ app/src/main/java/com/tecmilenio/booktrackerevidencia/MainActivity.kt
@@ -70,17 +70,9 @@ fun MainScreen() {
                 Spacer(modifier = Modifier.height(20.dp))
 
                 // Barra de búsqueda
-                OutlinedTextField(
-                    value = searchText,
-                    onValueChange = { searchText = it },
-                    placeholder = { Text("Buscar libro...") },
-                    modifier = Modifier.fillMaxWidth(),
-                    shape = RoundedCornerShape(20.dp),
-                    colors = TextFieldDefaults.outlinedTextFieldColors(
-                        containerColor = Color.White,
-                        focusedBorderColor = Color.Transparent,
-                        unfocusedBorderColor = Color.Transparent
-                    )
+                SearchBar(
+                    searchText = searchText,
+                    onSearchTextChanged = { searchText = it }
                 )
 
                 Spacer(modifier = Modifier.height(40.dp))
@@ -94,42 +86,7 @@ fun MainScreen() {
                     horizontalArrangement = Arrangement.spacedBy(10.dp)
                 ) {
                     items(filteredItems.size) { index ->
-                        val book = filteredItems[index] // Usar filteredItems en lugar de books
-
-                        Column(
-                            horizontalAlignment = Alignment.CenterHorizontally,
-                            modifier = Modifier
-                                .size(200.dp)
-                                .padding(2.dp)
-                        ) {
-                            Box(
-                                modifier = Modifier
-                                    .size(150.dp)
-                                    .background(Color(141, 141, 140), shape = RoundedCornerShape(5.dp))
-                                    .padding(5.dp)
-                            ) {
-                                if (book.imageUri != null) {
-                                    Image(
-                                        painter = rememberAsyncImagePainter(book.imageUri),
-                                        contentDescription = "Portada del libro",
-                                        modifier = Modifier.fillMaxSize()
-                                    )
-                                } else {
-                                    Box(
-                                        modifier = Modifier
-                                            .fillMaxSize() // Ahora ocupa todo el espacio disponible
-                                            .background(Color.Gray, shape = RoundedCornerShape(5.dp))
-                                    )
-                                }
-                            }
-
-                            Text(
-                                text = book.title,
-                                fontWeight = FontWeight.Bold,
-                                color = Color.White,
-                                modifier = Modifier.padding(top = 10.dp)
-                            )
-                        }
+                        BookCard(book = filteredItems[index]) // Usamos BookCard aquí
                     }
                 }
             }
@@ -158,112 +115,6 @@ fun MainScreen() {
     }
 }
 
-
-
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AddBookDialog(
-    onDismiss: () -> Unit,
-    onBookAdded: (Book) -> Unit
-) {
-    var title by remember { mutableStateOf("") }
-    var author by remember { mutableStateOf("") }
-    var totalPages by remember { mutableStateOf("") }
-    var genre by remember { mutableStateOf("") }
-    var startDate by remember { mutableStateOf("") }
-    var imageUri by remember { mutableStateOf<String?>(null) }
-
-    var errorMessage by remember { mutableStateOf<String?>(null) } // Mensaje de error
-
-    val context = LocalContext.current
-    val launcher = rememberLauncherForActivityResult(ActivityResultContracts.GetContent()) { uri ->
-        uri?.let { imageUri = it.toString() }
-    }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Agregar Libro") },
-        text = {
-            Column {
-                if (errorMessage != null) {
-                    Text(
-                        text = errorMessage!!,
-                        color = Color.Red,
-                        modifier = Modifier.padding(bottom = 10.dp)
-                    )
-                }
-
-                OutlinedTextField(
-                    value = title,
-                    onValueChange = { title = it },
-                    label = { Text("Título") },
-                    isError = title.isEmpty() && errorMessage != null // Resaltar campo vacío
-                )
-                OutlinedTextField(
-                    value = author,
-                    onValueChange = { author = it },
-                    label = { Text("Autor") },
-                    isError = author.isEmpty() && errorMessage != null
-                )
-                OutlinedTextField(
-                    value = totalPages,
-                    onValueChange = { totalPages = it },
-                    label = { Text("Páginas Totales") },
-                    isError = (totalPages.isEmpty() || totalPages.toIntOrNull() == null) && errorMessage != null
-                )
-                OutlinedTextField(
-                    value = genre,
-                    onValueChange = { genre = it },
-                    label = { Text("Género") },
-                    isError = genre.isEmpty() && errorMessage != null
-                )
-                OutlinedTextField(
-                    value = startDate,
-                    onValueChange = { startDate = it },
-                    label = { Text("Fecha de inicio") },
-                    isError = startDate.isEmpty() && errorMessage != null
-                )
-
-                Spacer(modifier = Modifier.height(10.dp))
-
-                Button(onClick = { launcher.launch("image/*") }) {
-                    Text("Seleccionar Imagen")
-                }
-
-                imageUri?.let {
-                    Image(
-                        painter = rememberAsyncImagePainter(it),
-                        contentDescription = "Imagen del libro",
-                        modifier = Modifier
-                            .size(100.dp)
-                            .padding(top = 10.dp)
-                    )
-                }
-            }
-        },
-        confirmButton = {
-            Button(onClick = {
-                if (title.isEmpty() || author.isEmpty() || totalPages.isEmpty() || genre.isEmpty() || startDate.isEmpty()) {
-                    errorMessage = "Todos los campos son obligatorios"
-                } else if (totalPages.toIntOrNull() == null) {
-                    errorMessage = "El número de páginas debe ser un valor válido"
-                } else {
-                    onBookAdded(Book(title, author, totalPages.toInt(), genre, startDate, imageUri))
-                    onDismiss()
-                }
-            }) {
-                Text("Agregar")
-            }
-        },
-        dismissButton = {
-            Button(onClick = onDismiss) {
-                Text("Cancelar")
-            }
-        }
-    )
-}
